Allow tab characters for search/find/replace
It is useful to be able to input tab characters into the prompts for
search and find/replace.  However, we cannot allow inputting tab
characters in the general case, since that would break tab completion.

Solve this by refactoring: move the contents of EditDynamicWithCallback
into EditDynamicWithCallbackAndTabs which has an option to allow tabs.
Update search and find/replace to use that new function.